% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{add}
\alias{add}
\alias{subtract}
\alias{multiply}
\alias{divide}
\alias{remainder}
\alias{mask}
\alias{maximum}
\alias{minimum}
\title{Binary operations}
\usage{
add(image, arg)

subtract(image, arg)

multiply(image, arg)

divide(image, arg)

remainder(image, arg)

mask(image, arg)

maximum(image, arg)

minimum(image, arg)
}
\arguments{
\item{image}{Image object or pipeline}

\item{arg}{Numeric or image argument}
}
\description{
Binary operations
}
