% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{drt_mean}
\alias{drt_mean}
\alias{drt_sd}
\alias{drt_max}
\alias{drt_whichmax}
\alias{drt_min}
\alias{drt_median}
\alias{drt_quantile}
\alias{drt_AR1}
\title{Dimensionality reduction in the temporal domain}
\usage{
drt_mean(image)

drt_sd(image)

drt_max(image)

drt_whichmax(image)

drt_min(image)

drt_median(image)

drt_quantile(image, prob)

drt_AR1(image)
}
\arguments{
\item{image}{An image object or pipeline.}

\item{prob}{For \code{drt_quantile}, the quantile probability to extract
(analogously to \code{\link{quantile}}).}
}
\value{
An updated pipeline.
}
\description{
Dimensionality reduction in the temporal domain
}
\examples{
image = system.file("extdata/example.nii.gz", package = "RNifti")
image \%>\% drt_mean \%>\% run()
drt_sd(image)
drt_max(image)
drt_whichmax(image)
drt_min(image)
drt_median(image)
drt_AR1(image)
}
