% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{imbibe}
\alias{imbibe}
\alias{asNifti.imbibe}
\alias{print.imbibe}
\title{Create an operation pipeline}
\usage{
imbibe(image)

\method{asNifti}{imbibe}(x, ...)

\method{print}{imbibe}(x, ...)
}
\arguments{
\item{image}{An image object or existing pipeline.}

\item{x}{An \code{"imbibe"} object.}

\item{...}{Additional arguments to \code{\link{run}}.}
}
\description{
Create an operation pipeline
}
