% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{kernel_3d}
\alias{kernel_3d}
\alias{kernel_2d}
\alias{kernel_box}
\alias{kernel_gauss}
\alias{kernel_sphere}
\alias{kernel_file}
\title{Mathematical morphology kernels}
\usage{
kernel_3d(image)

kernel_2d(image)

kernel_box(image, width, voxels = FALSE)

kernel_gauss(image, sigma)

kernel_sphere(image, radius)

kernel_file(image, file)
}
\description{
Mathematical morphology kernels
}
