% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{dilate}
\alias{dilate}
\alias{dilateall}
\alias{erode}
\alias{filter_median}
\alias{filter_mean}
\alias{smooth}
\title{Mathematical morphology and filtering operations}
\usage{
dilate(image, kernel = NULL, ..., max = FALSE, nonzero = TRUE)

dilateall(image, kernel = NULL, ...)

erode(image, kernel = NULL, ..., min = FALSE)

filter_median(image, kernel = NULL, ...)

filter_mean(image, kernel = NULL, ..., norm = TRUE)

smooth(image, sigma)
}
\description{
Mathematical morphology and filtering operations
}
