% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{dim_mean}
\alias{dim_mean}
\alias{dim_sd}
\alias{dim_max}
\alias{dim_whichmax}
\alias{dim_min}
\alias{dim_median}
\alias{dim_quantile}
\alias{dim_AR1}
\title{Dimensionality reduction operations}
\usage{
dim_mean(image, dim = 4L)

dim_sd(image, dim = 4L)

dim_max(image, dim = 4L)

dim_whichmax(image, dim = 4L)

dim_min(image, dim = 4L)

dim_median(image, dim = 4L)

dim_quantile(image, dim = 4L, prob)

dim_AR1(image, dim = 4L)
}
\arguments{
\item{image}{An image object or pipeline.}

\item{dim}{Integer value between 1 and 4, giving the dimension to apply the
reduction along.}

\item{prob}{For \code{drt_quantile}, the quantile probability to extract
(analogously to \code{\link{quantile}}).}
}
\value{
An updated pipeline.
}
\description{
Dimensionality reduction operations
}
