% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{run}
\alias{run}
\title{Run a pipeline and return an image result}
\usage{
run(pipe, precision = c("double", "float", "single"))
}
\arguments{
\item{pipe}{An operation pipeline.}

\item{precision}{The internal precision used for calculations. \code{"float"} and
\code{"single"} are equivalent.}
}
\value{
An image
}
\description{
Run a pipeline and return an image result
}
