% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{exponent}
\alias{exponent}
\alias{logarithm}
\alias{sine}
\alias{cosine}
\alias{tangent}
\alias{arcsine}
\alias{arccosine}
\alias{arctangent}
\alias{square}
\alias{squareroot}
\alias{reciprocal}
\alias{absolute}
\alias{binarise}
\title{Basic unary operations}
\usage{
exponent(image)

logarithm(image)

sine(image)

cosine(image)

tangent(image)

arcsine(image)

arccosine(image)

arctangent(image)

square(image)

squareroot(image)

reciprocal(image)

absolute(image)

binarise(image, invert = FALSE)
}
\description{
Basic unary operations
}
