% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{add}
\alias{add}
\alias{subtract}
\alias{multiply}
\alias{divide}
\alias{remainder}
\alias{mask}
\alias{maximum}
\alias{minimum}
\title{Basic binary operations}
\usage{
add(image, arg)

subtract(image, arg)

multiply(image, arg)

divide(image, arg)

remainder(image, arg)

mask(image, arg)

maximum(image, arg)

minimum(image, arg)
}
\arguments{
\item{image}{An image object or pipeline.}

\item{arg}{Numeric or image argument.}
}
\value{
An updated pipeline.
}
\description{
Basic binary operations
}
