% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run}
\alias{run}
\title{Run a pipeline and return an image result}
\usage{
run(pipe, precision = getOption("imbibe.precision", "double"))
}
\arguments{
\item{pipe}{An operation pipeline.}

\item{precision}{The internal precision used for calculations. May be
\code{"double"}, \code{"float"} or \code{"single"}; the latter two are
equivalent.}
}
\value{
An image
}
\description{
Run a pipeline and return an image result
}
\examples{
im <- RNifti::readNifti(system.file("extdata", "example.nii.gz", package="RNifti"))
pipe <- im \%>\% threshold_below(500) \%>\% binarise()
run(pipe)
}
