% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{exponent}
\alias{exponent}
\alias{logarithm}
\alias{sine}
\alias{cosine}
\alias{tangent}
\alias{arcsine}
\alias{arccosine}
\alias{arctangent}
\alias{square}
\alias{squareroot}
\alias{reciprocal}
\alias{absolute}
\alias{binarise}
\alias{binarize}
\title{Basic unary operations}
\usage{
exponent(image)

logarithm(image)

sine(image)

cosine(image)

tangent(image)

arcsine(image)

arccosine(image)

arctangent(image)

square(image)

squareroot(image)

reciprocal(image)

absolute(image)

binarise(image, invert = FALSE)

binarize(image, invert = FALSE)
}
\arguments{
\item{image}{An image object or pipeline.}

\item{invert}{Logical value: if \code{TRUE}, binarising will also perform
logical inversion so that only zeroes in the original image will be
nonzero; if \code{FALSE}, the default, the usual sense is used, in which
zeroes remain as they are, and everything else is converted to 1.}
}
\value{
An updated pipeline.
}
\description{
Basic unary operations
}
